requirements.txt
fastapi
gunicorn
joblib
numpy
pandas
prometheus-client
pydantic
pytest
python-dotenv
python-multipart
redis
requests
scikit-learn
streamlit
uvicorn[standard]
